HMP_2019_ibdmdb: IBDMDB TreeSummarizedExperiment data

An example gut microbiome dataset derived from the curatedMetagenomicData HMP_2019_ibdmdb resource (iHMP/HMP2 IBD cohort; Lloyd-Price et al., 2019).

data(HMP_2019_ibdmdb)

Format

A TreeSummarizedExperiment with 585 microbial species (rows) and 44 stool samples (columns) from 22 subjects with inflammatory bowel disease (two time points per subject: baseline T0 and follow-up T1).

Assays include relative abundance (relabundance) and a scaled version of the same data (scaled). Taxonomic annotation is stored in rowData, and a phylogenetic tree is available via rowTree. Sample metadata (colData) includes subject identifiers, time point, IBD subtype (CD/UC), and additional technical and demographic variables.

Source

Derived from curatedMetagenomicData (Lloyd-Price et al., 2019).

Details

The object is provided as a TreeSummarizedExperiment.

References

Lloyd-Price J, Arze C, Ananthakrishnan AN, et al. (2019). Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases. Nature 569, 655–662. doi:10.1038/s41586-019-1237-9

Examples

data("HMP_2019_ibdmdb", package = "miaViz")
HMP_2019_ibdmdb
#> class: TreeSummarizedExperiment 
#> dim: 585 44 
#> metadata(0):
#> assays(2): relabundance scaled
#> rownames(585): species:Phocaeicola vulgatus species:Bacteroides
#>   uniformis ... species:Enterococcus dispar species:Atopobium minutum
#> rowData names(7): superkingdom phylum ... genus species
#> colnames(44): CSM5FZ4E_P CSM5MCUQ_P ... PSM7J13I PSMA267H
#> colData names(27): study_name subject_id ... time_point treatment
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):
#> rowLinks: a LinkDataFrame (585 rows)
#> rowTree: 1 phylo tree(s) (10430 leaves)
#> colLinks: NULL
#> colTree: NULL
SummarizedExperiment::assayNames(HMP_2019_ibdmdb)
#> [1] "relabundance" "scaled"      
TreeSummarizedExperiment::rowTree(HMP_2019_ibdmdb)
#> 
#> Phylogenetic tree with 10430 tips and 10429 internal nodes.
#> 
#> Tip labels:
#>   k__Archaea|p__Candidatus_Micrarchaeota|c__Candidatus_Micrarchaeota_unclassified|o__Candidatus_Micrarchaeota_unclassified|f__Candidatus_Micrarchaeota_unclassified|g__Candidatus_Micrarchaeota_unclassified|s__Candidatus_Micrarchaeota_archaeon_CG1_02_55_22, k__Archaea|p__Archaea_unclassified|c__Archaea_unclassified|o__Archaea_unclassified|f__Archaea_unclassified|g__Archaea_unclassified|s__archaeon_GW2011_AR15, k__Archaea|p__Candidatus_Diapherotrites|c__Candidatus_Diapherotrites_unclassified|o__Candidatus_Diapherotrites_unclassified|f__Candidatus_Diapherotrites_unclassified|g__Candidatus_Diapherotrites_unclassified|s__Candidatus_Diapherotrites_archaeon_CG08_land_8_20_14_0_20_34_12, k__Archaea|p__Archaea_unclassified|c__Archaea_unclassified|o__Archaea_unclassified|f__Archaea_unclassified|g__Archaea_unclassified|s__archaeon_GW2011_AR10, k__Archaea|p__Candidatus_Diapherotrites|c__Candidatus_Diapherotrites_unclassified|o__Candidatus_Diapherotrites_unclassified|f__Candidatus_Diapherotrites_unclassified|g__Candidatus_Diapherotrites_unclassified|s__Candidatus_Diapherotrites_archaeon_CG11_big_fil_rev_8_21_14_0_20_37_9, k__Archaea|p__Euryarchaeota|c__Euryarchaeota_unclassified|o__Euryarchaeota_unclassified|f__Euryarchaeota_unclassified|g__Euryarchaeota_unclassified|s__Euryarchaeota_archaeon_TMED173, ...
#> 
#> Rooted; includes branch length(s).