Additional arguments for plotting

To be able to fine tune plotting, several additional plotting arguments are available. These are described on this page.

Tree plotting

line_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line_width:

Numeric scalar, specifying the default width of an edge. Defaults to NULL to use default of the ggtree package

line_width_range:

Two numeric values, the range for plotting dynamic edge widths in. Defaults to c(0.5,3).

point_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tips. Defaults to 1.

point_size:

Numeric scalar, specifying the default size of tips Defaults to 2..

point_size_range:

Two numeric values, the range for plotting dynamic tip sizes in. Defaults to c(1,4).

label_font_size:

Numeric scalar, font size for the tip and node labels. Default to 3.

highlight_font_size:

Numeric scalar, font size for the highlight labels. Default to 3.

Graph plotting

line_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line_width:

Numeric scalar, specifying the default width of an edge. Defaults to NULL to use default of the ggraph package

line_width_range:

Two numeric values, the range for plotting dynamic edge widths in. Defaults to c(0.5,3).

point_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tips. Defaults to 1.

point_size:

Numeric scalar, specifying the default size of tips Defaults to 2..

point_size_range:

Two numeric values, the range for plotting dynamic tip sizes in. Defaults to c(1,4).

Abundance plotting

flipped:

Logical scalar. Should the plot be flipped? Defaults to FALSE.

add_legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

add_x_text:

Logical scalar. Should x tick labels be plotted? Defaults to FALSE.

add_border:

Logical scalar. Should border of bars be plotted? Defaults to FALSE.

bar_alpha:

Numeric scalar in [0, 1], specifying the transparency of the bars. Defaults to 1.

point_alpha:

Numeric scalar in [0, 1], specifying the transparency of the points. Defaults to 1.

point_size:

Numeric scalar, specifying the default size of points. Defaults to 2..

Abundance density plotting

add_legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

point_shape:

Numeric scalar setting the shape of points. Defaults to 21.

point_colour:

Character scalar, specifying the default colour of points. Defaults to 2..

point_size:

Numeric scalar, specifying the default size of points. Defaults to 2..

point_alpha:

Numeric scalar in [0, 1], specifying the transparency of the points. Defaults to 1.

flipped:

Logical scalar. Should the plot be flipped? Defaults to FALSE.

scales_free:

Logical scalar. Should scales = "free" be set for faceted plots? Defaults to TRUE.

angle_x_text:

Logical scalar. Should x tick labels be plotted? Defaults to FALSE.

Prevalence plotting

flipped:

Logical scalar, specifying whether the plot should be flipped. Defaults to FALSE.

add_legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

point_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tips. Defaults to 1.

point_size:

Numeric scalar, specifying the default size of tips Defaults to 2..

line_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line_type:

Numeric scalar, specifying the default line type. Defaults to NULL to use default of the ggplot2 package

line_size:

Numeric scalar, specifying the default width of a line. Defaults to NULL to use default of the ggplot2 package

Series plotting

add_legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

line_alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line_type:

Numeric scalar, specifying the default line type. Defaults to NULL to use default of the ggplot2 package

line_width:

Numeric scalar, specifying the default width of a line. Defaults to NULL to use default of the ggplot2 package

line_width_range:

Two numeric values, the range for plotting dynamic line widths in. Defaults to c(0.5,3).

ribbon_alpha:

Numeric scalar in [0, 1], specifying the transparency of the ribbon. Defaults to 0.3.

Tile plotting

add_legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

rect_alpha:

Numeric scalar in [0, 1], specifying the transparency of the areas. Defaults to 1.

rect_colour:

Character scalar, specfiying the colour to use for colouring the borders of the areas. Defaults to "black".

na.value:

Character scalar, specfiying the colour to use for NA values. Defaults to "grey80".