Plot Series

This function plots series data.

plotSeries(
  object,
  x,
  y = NULL,
  rank = NULL,
  colour.by = colour_by,
  colour_by = NULL,
  size.by = size_by,
  size_by = NULL,
  linetype.by = linetype_by,
  linetype_by = NULL,
  assay.type = assay_name,
  assay_name = "counts",
  ...
)

# S4 method for class 'SummarizedExperiment'
plotSeries(
  object,
  x,
  y = NULL,
  rank = NULL,
  colour.by = colour_by,
  colour_by = NULL,
  size.by = size_by,
  size_by = NULL,
  linetype.by = linetype_by,
  linetype_by = NULL,
  assay.type = assay_name,
  assay_name = "counts",
  ...
)

Arguments

object

a SummarizedExperiment object.

x

a single character value for selecting the column from ColData that will specify values of x-axis.

y

a single character value for selecting the taxa from rownames. This parameter specifies taxa whose abundances will be plotted.

rank

a single character value defining a taxonomic rank, that is used to agglomerate the data. Must be a value of taxonomicRanks() function.

colour.by

a single character value defining a taxonomic rank, that is used to color plot. Must be a value of taxonomicRanks() function.

colour_by

Deprecated. Use colour.by instead.

size.by

a single character value defining a taxonomic rank, that is used to divide taxa to different line size types. Must be a value of taxonomicRanks() function.

size_by

Deprecated. Use size.by instead.

linetype.by

a single character value defining a taxonomic rank, that is used to divide taxa to different line types. Must be a value of taxonomicRanks() function.

linetype_by

Deprecated. Use linetype.by instead.

assay.type

a single character value for selecting the assay to be plotted. (default: assay.type = "counts")

assay_name

a single character value for specifying which assay to use for calculation. (Please use assay.type instead. At some point assay_name will be disabled.)

...

additional parameters for plotting. See mia-plot-args for more details i.e. call help("mia-plot-args")

Value

A ggplot2 object

Details

This function creates series plot, where x-axis includes e.g. time points, and y-axis abundances of selected taxa.

Author

Leo Lahti and Tuomas Borman. Contact: microbiome.github.io

Examples

if (FALSE) { # \dontrun{
library(mia)
# Load data from miaTime package
library("miaTime")
data(SilvermanAGutData)
object <- SilvermanAGutData

# Plots 2 most abundant taxa, which are colored by their family
plotSeries(object,
           x = "DAY_ORDER",
           y = getTop(object, 2),
           colour.by = "Family")

# Counts relative abundances
object <- transformAssay(object, method = "relabundance")

# Selects taxa
taxa <- c("seq_1", "seq_2", "seq_3", "seq_4", "seq_5")

# Plots relative abundances of phylums
plotSeries(object[taxa,],
           x = "DAY_ORDER", 
           colour.by = "Family",
           linetype.by = "Phylum",
           assay.type = "relabundance")

# In addition to 'colour.by' and 'linetype.by', 'size.by' can also be used to group taxa.
plotSeries(object,
           x = "DAY_ORDER", 
           y = getTop(object, 5), 
           colour.by = "Family",
           size.by = "Phylum",
           assay.type = "counts")
} # }