Additional arguments for plotting

To be able to fine tune plotting, several additional plotting arguments are available. These are described on this page.

Tree plotting

line.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line.width:

Numeric scalar, specifying the default width of an edge. Defaults to NULL to use default of the ggtree package

line.width.range:

Two numeric values, the range for plotting dynamic edge widths in. Defaults to c(0.5,3).

point.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tips. Defaults to 1.

point.size:

Numeric scalar, specifying the default size of tips Defaults to 2..

point.size.range:

Two numeric values, the range for plotting dynamic tip sizes in. Defaults to c(1,4).

label.font.size:

Numeric scalar, font size for the tip and node labels. Default to 3.

highlight.font.size:

Numeric scalar, font size for the highlight labels. Default to 3.

Graph plotting

line.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line.width:

Numeric scalar, specifying the default width of an edge. Defaults to NULL to use default of the ggraph package

line.width.range:

Two numeric values, the range for plotting dynamic edge widths in. Defaults to c(0.5,3).

point.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tips. Defaults to 1.

point.size:

Numeric scalar, specifying the default size of tips Defaults to 2..

point.size.range:

Two numeric values, the range for plotting dynamic tip sizes in. Defaults to c(1,4).

Abundance plotting

flipped:

Logical scalar. Should the plot be flipped? Defaults to FALSE.

add.legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

add.x.text:

Logical scalar. Should x tick labels be plotted? Defaults to FALSE.

add.border:

Logical scalar. Should border of bars be plotted? Defaults to FALSE.

bar.alpha:

Numeric scalar in [0, 1], specifying the transparency of the bars. Defaults to 1.

point.alpha:

Numeric scalar in [0, 1], specifying the transparency of the points. Defaults to 1.

point.size:

Numeric scalar, specifying the default size of points. Defaults to 2..

Abundance density plotting

add.legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

point.shape:

Numeric scalar setting the shape of points. Defaults to 21.

point.colour:

Character scalar, specifying the default colour of points. Defaults to 2..

point.size:

Numeric scalar, specifying the default size of points. Defaults to 2..

point.alpha:

Numeric scalar in [0, 1], specifying the transparency of the points. Defaults to 1.

flipped:

Logical scalar. Should the plot be flipped? Defaults to FALSE.

scales.free:

Logical scalar. Should scales = "free" be set for faceted plots? Defaults to TRUE.

angle.x.text:

Logical scalar. Should x tick labels be plotted? Defaults to FALSE.

Prevalence plotting

flipped:

Logical scalar, specifying whether the plot should be flipped. Defaults to FALSE.

add.legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

point.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tips. Defaults to 1.

point.size:

Numeric scalar, specifying the default size of tips Defaults to 2..

line.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line.type:

Numeric scalar, specifying the default line type. Defaults to NULL to use default of the ggplot2 package

line.size:

Numeric scalar, specifying the default width of a line. Defaults to NULL to use default of the ggplot2 package

Series plotting

add.legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

line.alpha:

Numeric scalar in [0, 1], specifying the transparency of the tree edges. Defaults to 1.

line.type:

Numeric scalar, specifying the default line type. Defaults to NULL to use default of the ggplot2 package

line.width:

Numeric scalar, specifying the default width of a line. Defaults to NULL to use default of the ggplot2 package

line.width.range:

Two numeric values, the range for plotting dynamic line widths in. Defaults to c(0.5,3).

ribbon.alpha:

Numeric scalar in [0, 1], specifying the transparency of the ribbon. Defaults to 0.3.

Tile plotting

add.legend:

Logical scalar. Should legends be plotted? Defaults to TRUE.

rect.alpha:

Numeric scalar in [0, 1], specifying the transparency of the areas. Defaults to 1.

rect.colour:

Character scalar, specifying the colour to use for colouring the borders of the areas. Defaults to "black".

na.value:

Character scalar, specifying the colour to use for NA values. Defaults to "grey80".