Additional arguments for plotting

To be able to fine tune plotting, several additional plotting arguments are available. These are described on this page.

Tree plotting

line.alpha:

Numeric scalar in [0, 1], Specifies the transparency of the tree edges. (Default: 1)

line.width:

Numeric scalar. Specifies the default width of an edge. (Default: NULL) to use default of the ggtree package.

line.width.range:

Numeric vector. The range for plotting dynamic edge widths in. (Default: c(0.5,3))

point.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the tips. (Defaults: 1)

point.size:

Numeric scalar. Specifies the default size of tips. (Defaults: 2.)

point.size.range:

Numeric vector. Specifies the range for plotting dynamic tip sizes in. (Defaults: c(1,4))

label.font.size:

Numeric scalar. Font size for the tip and node labels. (Default: 3)

highlight.font.size:

Numeric scalar. Font size for the highlight labels. (Default: 3)

Graph plotting

line.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the tree edges. (Default: 1)

line.width:

Numeric scalar. Specifies the default width of an edge. (Default: NULL) to use default of the ggtree package.

line.width.range:

Numeric vector. The range for plotting dynamic edge widths in. (Default: c(0.5,3))

point.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the tips. (Default: 1)

point.size:

Numeric scalar. Specifies the default size of tips. (Default: 2.)

point.size.range:

Numeric vector. The range for plotting dynamic tip sizes in. (Default: c(1,4))

Abundance plotting

flipped:

Logical scalar. Should the plot be flipped? (Default: FALSE)

add.legend:

Logical scalar. Should legends be plotted? (Default: TRUE)

add.x.text:

Logical scalar. Should x tick labels be plotted? (Default: FALSE)

add.border:

Logical scalar. Should border of bars be plotted? (Default: FALSE)

bar.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the bars. (Default: 1)

point.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the points. (Default: 1)

point.size:

Numeric scalar. Specifies the default size of points. (Default: 2.)

Abundance density plotting

add.legend:

Logical scalar. Should legends be plotted? (Defaults: TRUE)

point.shape:

Numeric scalar. Sets the shape of points. (Default: 21)

point.colour:

Character scalar. Specifies the default colour of points. (Default: 2.)

point.size:

Numeric scalar. Specifies the default size of points. (Default: 2.)

point.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the points. (Default: 1)

flipped:

Logical scalar. Should the plot be flipped? (Default: FALSE)

scales.free:

Logical scalar. Should scales = "free" be set for faceted plots? (Default: TRUE)

angle.x.text:

Logical scalar. Should x tick labels be plotted? (Default: FALSE)

Prevalence plotting

flipped:

Logical scalar. Specifies whether the plot should be flipped. (Default: FALSE)

add.legend:

Logical scalar. Should legends be plotted? (Default: TRUE)

point.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the tips. (Default: 1)

point.size:

Numeric scalar. Specifies the default size of tips. (Default: 2.)

line.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the tree edges. (Default: 1)

line.type:

Numeric scalar. Specifies the default line type. (Default: NULL) to use default of the ggplot2 package.

line.size:

Numeric scalar. Specifies the default width of a line. (Default: NULL) to use default of the ggplot2 package.

Series plotting

add.legend:

Logical scalar. Should legends be plotted? (Default: TRUE)

line.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the tree edges. (Default: 1)

line.type:

Numeric scalar. Specifies the default line type. (Default: NULL) to use default of the ggplot2 package.

line.width:

Numeric scalar. Specifies the default width of a line. (Default: NULL) to use default of the ggplot2 package.

line.width.range:

Numeric vector. The range for plotting dynamic line widths in. (Default: c(0.5,3))

ribbon.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the ribbon. (Default: 0.3)

Tile plotting

add.legend:

Logical scalar. Should legends be plotted? (Default: TRUE)

rect.alpha:

Numeric scalar in [0, 1]. Specifies the transparency of the areas. (Default: 1)

rect.colour:

Character scalar. Specifies the colour to use for colouring the borders of the areas. (Default: "black")

na.value:

Character scalar. Specifies the colour to use for NA values. (Default: "grey80")